+++ src/main/java/pl/betoncraft/betonquest/config/ConfigUpdater.java
@@ -343,6 +343,9 @@ public class ConfigUpdater {
     private void updateFromV53() {
         final ConfigurationSection section = config.getConfigurationSection("effectlib_npc_effect");
         if (section != null) {
+            if (Config.getDefaultPackage() == null) {
+                Config.createDefaultPackage(Optional.ofNullable(config.getString("default_package")).orElse("default"));
+            }
             final ConfigAccessor custom = Config.getDefaultPackage().getCustom();
             final Configuration config = custom.getConfig();
             config.set("npc_effects.default", section);